Refactor CartCounter component and counterSlice
This commit refactors the CartCounter component and counterSlice in the shopping cart module. The `resetCount` action in the counterSlice has been replaced with the `initCounterState` action, which initializes the counter state with a given value. The `isReady` property has been added to the counter state to track the initialization status.